requirements.txt
# Core Python & Utilities
numpy==1.26.4
pandas==2.1.4
python-dotenv==1.0.1

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0

# Machine Learning / Deep Learning
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.4.0

# API Development
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Frontend (UI)
streamlit==1.30.0
requests==2.31.0

# Logging & Monitoring
loguru==0.7.2

# Data Visualization (optional but useful)
matplotlib==3.8.2
seaborn==0.13.1

# Development & Formatting (optional)
black==23.12.1
isort==5.13.2

jupyterlab==4.3.1
ipykernel
tqdm==4.66.4